refactor: decoupled map components to maintain orthoganlity
